+++ src/api/Api.jsx
@@ -1,17 +1,5 @@
 import * as Mock from "./mocked-data.jsx";
 
-export function apiCallToGetEmployeeMetric(employeeId) {
-  return Mock.Employees[0];
-}
-
-export function apiCallToGetEmployeeOffice(officeId) {
-  return Mock.Office;
-}
-
-export function apiCallToGetEmployeeWorkstation(workstationId) {
-  return Mock.Workstation;
-}
-
 export function apiCallToGetEmployeeCarbonFootprint(employeeId) {
   return Mock.Carbon;
 }
@@ -28,7 +16,7 @@ export async function apiCallToGetCarbonThresholds() {
 }
 
 export async function apiCallToGetCompanies() {
-  const response = await fetch('http://localhost:8080/company/companies');
+  const response = await fetch('http://localhost:8080/company/companies?isSimple=true');
   const resData = await response.json();
 
   if (!response.ok) {
@@ -39,7 +27,7 @@ export async function apiCallToGetCompanies() {
 }
 
 export async function apiCallToGetAreas(companyId, companyName) {
-  const response = await fetch('http://localhost:8080/area/areas?id='+companyId+'&company='+companyName);
+  const response = await fetch('http://localhost:8080/area/areas?id='+companyId+'&company='+companyName+'&isSimple=true');
   const resData = await response.json();
 
   if (!response.ok) {
@@ -50,7 +38,7 @@ export async function apiCallToGetAreas(companyId, companyName) {
 }
 
 export async function apiCallToGetTribes(areaId, areaName) {
-  const response = await fetch('http://localhost:8080/tribe/tribes?id='+areaId+'&area='+areaName);
+  const response = await fetch('http://localhost:8080/tribe/tribes?id='+areaId+'&area='+areaName+'&isSimple=true');
   const resData = await response.json();
 
   if (!response.ok) {
@@ -61,7 +49,7 @@ export async function apiCallToGetTribes(areaId, areaName) {
 }
 
 export async function apiCallToGetTeams(tribeId, tribeName) {
-  const response = await fetch('http://localhost:8080/team/teams?id='+tribeId+'&tribe='+tribeName);
+  const response = await fetch('http://localhost:8080/team/teams?id='+tribeId+'&tribe='+tribeName+'&isSimple=true');
   const resData = await response.json();
 
   if (!response.ok) {
@@ -72,11 +60,22 @@ export async function apiCallToGetTeams(tribeId, tribeName) {
 }
 
 export async function apiCallToGetEmployees(teamId, teamName) {
-  const response = await fetch('http://localhost:8080/employee/allByTeam?id='+teamId+'&team='+teamName);
+  const response = await fetch('http://localhost:8080/employee/allByTeam?id='+teamId+'&team='+teamName+'&isSimple=true');
   const resData = await response.json();
 
   if (!response.ok) {
     throw new Error("Failed to get employees.");
+  }
+
+  return resData;
+}
+
+export async function apiCallToGetEmployeeMetric(employeeId, corporateKey) {
+  const response = await fetch('http://localhost:8080/employee/employee?id='+employeeId+'&ck='+corporateKey);
+  const resData = await response.json();
+
+  if (!response.ok) {
+    throw new Error("Failed to get employee.");
   }
 
   return resData;