+++ src/components/Main/Metrics/MetricsEmployee/Office/Office.jsx
@@ -1,6 +1,5 @@
 import { memo, useEffect, useState } from "react";
 import { RiBuilding2Line } from "react-icons/ri";
-import { apiCallToGetEmployeeOffice } from "../../../../../api/Api.jsx";
 import {
   $Content,
   $Head,
@@ -10,24 +9,19 @@ import {
   $Title,
 } from "./Office.styles.jsx";
 
-const Office = memo(function Office({ office_id }) {
+const Office = memo(function Office({ office }) {
   const [employeeOffice, setEmployeeOffice] = useState({});
 
   useEffect(() => {
-    const office = apiCallToGetEmployeeOffice(office_id);
-    setEmployeeOffice((_prevOffice) => office);
-  }, [office_id]);
+    const { name, street, streetNumber, apartmentNumber, postalCode, location: ulocation } = office;
+    const location = ulocation ? getLocationData(ulocation) : '';
+    setEmployeeOffice((_prevOffice) => ({name, street, streetNumber, apartmentNumber, postalCode, ...location}));
+  }, [office]);
 
-  const { apartment_number, name, street, street_number, location } = employeeOffice;
-  let ecity;
-  let ecountry;
-
-  if (location) {
-    const { city, country } = location;
-    ecity = city;
-    if (country) {
-      ecountry = country.name;
-    }
+  function getLocationData(data) {
+    const { city, country: ucountry } = data;
+    const location = { city, country: ucountry && ucountry.name};
+    return location;
   }
 
   return (
@@ -36,31 +30,36 @@ const Office = memo(function Office({ office_id }) {
         <$Icon>
           <RiBuilding2Line />
         </$Icon>
-        <$Title>{name}</$Title>
+        <$Title>{employeeOffice.name}</$Title>
       </$Head>
       <$Content>
         <div>Street:</div>
-        <div>{street}</div>
+        <div>{employeeOffice.street}</div>
       </$Content>
       <$Line />
       <$Content>
         <div>Street number:</div>
-        <div>{street_number}</div>
+        <div>{employeeOffice.streetNumber}</div>
       </$Content>
       <$Line />
       <$Content>
         <div>Apartment number:</div>
-        <div>{apartment_number}</div>
+        <div>{employeeOffice.apartmentNumber !== 0 ? employeeOffice.apartmentNumber : "n/a"}</div>
+      </$Content>
+      <$Line />
+      <$Content>
+        <div>Postal code:</div>
+        <div>{employeeOffice.postalCode}</div>
       </$Content>
       <$Line />
       <$Content>
         <div>City:</div>
-        <div>{ecity}</div>
+        <div>{employeeOffice.city}</div>
       </$Content>
       <$Line />
       <$Content>
         <div>Country:</div>
-        <div>{ecountry}</div>
+        <div>{employeeOffice.country}</div>
       </$Content>
     </$Office>
   );